sr(api): initialize a legalDocuments context for builders
