+++ api/tests/deprecated/unit/infrastructure/serializers/jsonapi/certification-point-of-contact.serializer.test.js
@@ -51,7 +51,7 @@ describe('Unit | Deprecated | Serializer | JSONAPI | certification-point-of-cont
       ];
       const lastPixCertifTermsOfServiceValidatedAt = new Date('2024-06-04');
 
-      const pixCertifTermsOfServiceStatus = domainBuilder.buildLegalDocumentStatus({
+      const pixCertifTermsOfServiceStatus = domainBuilder.legalDocuments.buildLegalDocumentStatus({
         status: STATUS.ACCEPTED,
         acceptedAt: lastPixCertifTermsOfServiceValidatedAt,
         documentPath: 'pix-certif-tos-2024-06-03',
