+++ api/tests/tooling/domain-builder/factory/legal-documents/build-legal-document-status.js
@@ -1,4 +1,5 @@
-import { LegalDocumentStatus, STATUS } from '../../../../src/legal-documents/domain/models/LegalDocumentStatus.js';
+import { LegalDocumentStatus, STATUS } from '../../../../../src/legal-documents/domain/models/LegalDocumentStatus.js';
+
 const buildLegalDocumentStatus = function ({ status = STATUS.ACCEPTED, acceptedAt = null, documentPath = null } = {}) {
   return new LegalDocumentStatus({
     status,
